+++ application/src/main/java/common/bolts/transactional/ob/OBBolt.java
@@ -44,7 +44,7 @@ public abstract class OBBolt extends TransactionalBolt {
     }
 
     @Override
-    protected void TXN_PROCESS(long _bid) throws DatabaseException, InterruptedException {
+    protected void TXN_PROCESS(double _bid) throws DatabaseException, InterruptedException {
     }
 
     protected void BUYING_REQUEST_LOCKAHEAD(BuyingEvent event, TxnContext txnContext) throws DatabaseException {
@@ -182,9 +182,9 @@ public abstract class OBBolt extends TransactionalBolt {
     }
 
     @Override
-    protected void POST_PROCESS(long _bid, long timestamp, int combo_bid_size) throws InterruptedException {
+    protected void POST_PROCESS(double _bid, long timestamp, int combo_bid_size) throws InterruptedException {
         BEGIN_POST_TIME_MEASURE(thread_Id);
-        for (long i = _bid; i < _bid + combo_bid_size; i++) {
+        for (double i = _bid; i < _bid + combo_bid_size; i++) {
             TxnEvent event = (TxnEvent) input_event;
             (event).setTimestamp(timestamp);
             if (event instanceof BuyingEvent) {
